Created schema.sql and seed.sql.
Created schema.sql to generate the three tables needed that are outlined in the module description. Created seed.sql. to populate the database and tables with mock data for testing.

Added a call to seed.sql within the schema file to auto populate the data and call a simple join to view the tables.